requirements.txt
Flask==3.0.2
gunicorn==21.2.0
Werkzeug==3.0.1
mysql-connector-python==8.3.0 